Create tsunami module
